Move ExportListener to its own file
